adding ssi mockup main
